+++ src/Tar/Encode.php
@@ -9,7 +9,10 @@ use Innmind\Filesystem\{
     Directory,
 };
 use Innmind\MediaType\MediaType;
-use Innmind\TimeContinuum\Clock;
+use Innmind\TimeContinuum\{
+    Clock,
+    Format,
+};
 use Innmind\Immutable\{
     Str,
     Sequence,
@@ -157,7 +160,7 @@ final class Encode
                 \sprintf('%07s', \decoct(0)), // user id
                 \sprintf('%07s', \decoct(0)), // group id
                 \sprintf('%011s', \decoct($size)), // file size
-                \sprintf('%011s', \decoct((int) ($this->clock->now()->milliseconds() / 1000))), // file last modification time
+                \sprintf('%011s', \decoct((int) $this->clock->now()->format(Format::of('U')))), // file last modification time
             ),
             Str\Encoding::ascii,
         );